# server/openapi.yaml
openapi: 3.0.0
paths:
...
  /agents/create:
    post:
      operationId: createAgent
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedAgentDto"
      tags:
        - agents
      security:
        - bearer: []
  /agents/uploadLogo/{agentId}:
    post:
      operationId: uploadLogo
      parameters:
        - name: agentId
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - agents
      security:
        - bearer: []
  /agents/update/{agentId}:
    post:
      operationId: updateAgent
      parameters:
        - name: agentId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedAgentDto"
        "401":
          description: You are not authorized to access this resource
      tags:
        - agents
      security:
        - bearer: []
  /agents/delete/{agentId}:
    delete:
      operationId: deleteAgent
      parameters:
        - name: agentId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedAgentResponseDto"
        "401":
          description: You are not authorized to access this resource
      tags:
        - agents
      security:
        - bearer: []
  /agents/listForProject/{projectId}:
    post:
      operationId: listForProject
      parameters:
        - name: projectId
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAgentsResponseDto"
        "401":
          description: You are not authorized to access this resource
      tags:
        - agents
      security:
        - bearer: []
  /agents/getById/{agentId}:
    post:
      operationId: getById
      parameters:
        - name: agentId
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAgentResponseDto"
        "401":
          description: You are not authorized to access this resource
      tags:
        - agents
      security:
        - bearer: []
...
info:
  title: AgentLabs API
  description: Build AI Agents in minutes, not months.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateAgentDto:
      type: object
      properties:
        name:
          type: string
        projectId:
          type: string
      required:
        - name
        - projectId
    CreatedAgentDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        projectId:
          type: string
        creatorId:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - name
        - projectId
        - creatorId
        - createdAt
        - updatedAt
    UpdateAgentDto:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UpdatedAgentDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        projectId:
          type: string
        creatorId:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - name
        - projectId
        - creatorId
        - createdAt
        - updatedAt
    DeletedAgentResponseDto:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    ListAgentItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        projectId:
          type: string
        creatorId:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - name
        - projectId
        - creatorId
        - createdAt
        - updatedAt
    ListAgentsResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ListAgentItem"
        total:
          type: number
      required:
        - items
        - total
    GetAgentResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        projectId:
          type: string
        creatorId:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - name
        - projectId
        - creatorId
        - createdAt
        - updatedAt